Fixes #1811 Remove duplicated Abstract Classes that we could use them from AppBundle
